% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_transcript.R
\name{build_dotplot}
\alias{build_dotplot}
\title{Custom GGPlot Function}
\usage{
build_dotplot(df, y_labs)
}
\arguments{
\item{df}{A dataframe containing the necessary columns.}

\item{y_labs}{A vector of \code{cds_id} values to filter the dataframe.}
}
\value{
A GGPlot object.
}
\description{
This function takes a dataframe and a vector of \code{cds_id} values, and creates a customized GGPlot using ggplot2.
}
