% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_transcript.R
\name{build_transcript_plot}
\alias{build_transcript_plot}
\title{Create Transcript Plot}
\usage{
build_transcript_plot(gtf)
}
\arguments{
\item{gtf}{A dataframe containing GTF information.}
}
\value{
A GGPlot object.
}
\description{
This function takes a dataframe containing GTF information and creates a
customized transcript plot using ggplot2.
}
\examples{
gtf <- data.frame(
  type = c("exon", "CDS", "exon", "exon", "CDS", "exon"),
  start = c(100, 150, 200, 250, 300, 350),
  end = c(130, 170, 220, 260, 310, 360),
  Name = c("transcript1", "transcript1", "transcript2", "transcript3", "transcript2", "transcript4"),
  cds_source2 = c("source1", "source1", "source2", "source2", "source3", "source3"),
  PTC = c(TRUE, FALSE, TRUE, FALSE, TRUE, FALSE),
  strand = c("+", "-", "+", "+", "-", "+")
)
build_transcript_plot(gtf)

}
