% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{create_grid}
\alias{create_grid}
\title{Simplifies grid creation}
\usage{
create_grid(areas, cols_width, rows_height)
}
\description{
Simplifies grid creation
}
\note{
does not support mobile grid.
}
