% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{create_trackhub_url}
\alias{create_trackhub_url}
\title{Create Trackhub URL}
\usage{
create_trackhub_url(
  base_url = "http://genome-euro.ucsc.edu/",
  db = "hg38",
  position = NA,
  hub =
    "https://trackhub.dieterichlab.org/tbrittoborges/nmd_transcriptome/nmd_transcriptome.hub.txt"
)
}
\arguments{
\item{base_url}{The base URL of the UCSC Genome Browser. Default is "http://genome-euro.ucsc.edu/".}

\item{db}{The genome database. Default is "hg38".}

\item{position}{The genomic position. Default is NA for no posiitoj.}

\item{hub}{The URL of the track hub. Default is "https://trackhub.dieterichlab.org/tbrittoborges/nmd_transcriptome/nmd_transcriptome.hub.txt".}
}
\value{
A URL for accessing the specified track hub on the UCSC Genome Browser.
}
\description{
This function generates a URL for accessing a track hub on the UCSC Genome Browser.
}
\examples{
create_trackhub_url()
create_trackhub_url(db = "mm10", position = "chr1:1000-2000")
create_trackhub_url(db = "mm10", position = "chr1:1000-2000", hub=NA)
}
