% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{fc_boxplot}
\alias{fc_boxplot}
\title{Plot Boxplot with Gene Highlight}
\usage{
fc_boxplot(data, index, gene_l2fc, xlims)
}
\arguments{
\item{data}{A data frame containing the necessary columns: gene_name, contrasts, log2FoldChange.}

\item{xlims}{minimum, maximum values for x-axis}

\item{gene_name}{The name of the gene to highlight.}
}
\value{
A ggplot object displaying the boxplot with the highlighted gene.
}
\description{
This function generates a boxplot with a highlighted gene based on the specified contrast.
}
