% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{make_unique}
\alias{make_unique}
\title{Make Unique Character Vectors with Suffixes}
\usage{
make_unique(strings)
}
\arguments{
\item{strings}{A character vector with potentially non-unique elements.}
}
\value{
A character vector where each element has been made unique by appending
an underscore and an occurrence number.
}
\description{
Takes a character vector and makes each element unique by appending an increment
}
\examples{
original_vector <- c("apple", "banana", "apple", "orange", "banana", "banana")
make_unique(original_vector)
}
