% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{plot_annotation_cdna}
\alias{plot_annotation_cdna}
\title{Plot a transcript on the cDNA coordiantes highlighting PTC containing
transcripts}
\usage{
plot_annotation_cdna(bed12, y_labs)
}
\description{
Plot a transcript on the cDNA coordiantes highlighting PTC containing
transcripts
}
