% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{position_from_gtf}
\alias{position_from_gtf}
\title{Extract Genomic Position from GTF}
\usage{
position_from_gtf(gtf, col = "gene_id")
}
\arguments{
\item{gtf}{A data frame containing the GTF data.}

\item{col}{The column to group by for summarization.}
}
\value{
A character string representing the genomic position in the format "chr{seqnames}:{start}-{end}".
}
\description{
This function extracts the genomic position from a GTF file by grouping the data based on a specified column and summarizing the sequence name, start, and end coordinates.
}
\examples{
gtf <- data.frame(seqnames = c("chr1", "chr1", "chr2", "chr2"),
                  start = c(100, 200, 300, 400),
                  end = c(500, 600, 700, 800),
                  gene_id = c("Gene1", "Gene2", "Gene3", "Gene4"))
position_from_gtf(gtf, "gene_id")

}
